requirements.txt
discord.py
python.dotenv
termcolor
loguru
